+++ src/Hobbit.java
@@ -7,8 +7,8 @@ public class Hobbit extends LivingBeing{
     };
     private static final int MIN_AGE = 18;
     private static final int MAX_AGE = 100;
-    private static final int MIN_WEIGHT = 50;
-    private static final int MAX_WEIGHT = 100;
+    private static final double MIN_WEIGHT = 50;
+    private static final double MAX_WEIGHT = 100;
 
     Hobbit(String name, int age, double weight) {
         super (name, age, weight);
